refactor(combo): extract combo structure resolution to combo/comboStructure.ts
