+++ open-sse/services/combo/comboData.ts
@@ -8,6 +8,17 @@
  * (no cycles). Logic unchanged; combo.ts imports them back for compatibility.
  */
 
+import type { ResolvedComboTarget } from "./types.ts";
+
 export function isRecord(value: unknown): value is Record<string, unknown> {
   return !!value && typeof value === "object" && !Array.isArray(value);
+}
+
+export function dedupeTargetsByExecutionKey(targets: ResolvedComboTarget[]) {
+  const seen = new Set<string>();
+  return targets.filter((target) => {
+    if (seen.has(target.executionKey)) return false;
+    seen.add(target.executionKey);
+    return true;
+  });
 }
